+++ src/OpenTelemetry.Instrumentation.SqlClient/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 * Instrumentation modified to depend only on the API.
+* Activities are now created with the `db.system` attribute set for usage
+  during sampling. ([#1979](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1979))
 
 ## 1.0.0-rc3